// src/models/enum/application.status.enum.ts
export enum ApplicationStatus {
  WAITING = "Chờ tiếp nhận",
  PRIORITY = "Diện ưu tiên",
  MEETS_CRITERIA = "Hồ sơ đạt tiêu chí",
  DOES_NOT_MEET_CRITERIA = "Không đạt tiêu chí",
  ACCEPTED = "Trúng tuyển",
  NOT_ACCEPTED = "Không trúng tuyển",
  ENROLLED = "Nhập học",
}